+++ lib/spinning_wheel.dart
@@ -5,5 +5,6 @@
 library spinning_wheel;
 
 export 'models/wheel_segment.dart';
+export 'models/wheel_label_style.dart';
 export 'spinner_wheel.dart';
 export 'controller/spin_controller.dart';